chore: rename element to chemicalElement
