+++ src/locales/en/science/chemicalElement.ts
